Fix propagating changes within jail wrapper
